refactor: Extract actions service to own module
